requirements.txt
botasaurus==4.*
gspread
google-auth
python-dotenv
supabase==2.*
flask
pandas
pywhatkit
playwright
tqdm